blueprint for BMS driver

+++ BMS_Playground/LIB/STD_Types.h
@@ -5,8 +5,8 @@
 */
 typedef unsigned char u8;
 typedef unsigned short int u16;
-typedef unsigned long int u32;
-typedef unsigned long long int u64;
+typedef unsigned int u32;
+typedef unsigned long int u64;
 
 /**
 	signed integers
